backend/requirement.txt
fastapi[all]
psycopg2-binary
uvicorn
python-dotenv
sqlalchemy
pydantic